moving runtime to subcommands dir
